Make it easier to place buildings on mobile

+++ jsettlers.logic/src/test/java/jsettlers/algorithms/construction/ConstructionMarksAlgorithmTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015
+ * Copyright (c) 2015 - 2017
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
@@ -196,7 +196,7 @@ public class ConstructionMarksAlgorithmTest {
 		}
 
 		@Override
-		public boolean canConstructAt(short x, short y, EBuildingType type, byte playerId) {
+		public boolean canConstructAt(int x, int y, EBuildingType type, byte playerId) {
 			throw new UnsupportedOperationException("not mocked");
 		}
